refactor: move nested routing logic up to parent directory
